+++ package.json
@@ -18,7 +18,8 @@
     "swiper": "^4.4.5",
     "velocity-animate": "^1.5.2",
     "vue": "^2.5.2",
-    "vue-router": "^3.0.1"
+    "vue-router": "^3.0.1",
+    "vuex": "^3.0.1"
   },
   "devDependencies": {
     "autoprefixer": "^7.1.2",
