+++ src/components/Foot.vue
@@ -25,12 +25,22 @@ let navConfig =[{
 }]
 
 import qs from 'qs'
+import bus from 'js/bus.js'
 export default {
+  props:['obj'],
   data() {
     return {
       navConfig,
-      curIndex: parseInt(qs.parse(location.search.substr(1)).index) || 0
+      curIndex: parseInt(qs.parse(location.search.substr(1)).index) || 0,
+      ob: JSON.parse(JSON.stringify(this.obj))
     }
+  },
+  created(){
+    setTimeout(() => {
+      this.ob.age = 18
+      // this.$emit('change',18)
+      bus.$emit('change',18)
+    },5000)
   },
   methods: {
     changeNav(list,index) {
