+++ src/pages/index/index.html
@@ -85,7 +85,7 @@
       <div class="js-show-find category-guid" style="display: none;"></div>
     </div>
   </div>
-    <foot></foot>
+    <foot :obj="obj" @change="changeAge"></foot>
   </div>
 </body>
