+++ src/pages/member/member.js
@@ -1,34 +1,9 @@
 
 //1.使用vue-router
+
+
 import Vue from 'vue'
-import Router from 'vue-router'
-
-
-Vue.use(Router)
-
-let routes = [{
-  path: '/',
-  component: require('./components/member.vue').default
-},{
-  path: '/address',
-  component: require('./components/address.vue').default,
-  children: [{
-    path: '',
-    redirect: 'all'
-  },{
-    path: 'all',
-    name: 'all',
-    component: require('./components/all.vue').default
-  },{
-    path: 'form',
-    name: 'form',
-    component: require('./components/form.vue').default
-  }]
-}]
-
-let router = new Router({
-  routes
-})
+import router from './router/index.js'
 
 new Vue({
   el: '#app',
