+++ src/pages/member/vuex/index.js
@@ -1,0 +1,14 @@
+import Vue from 'vue'
+import Vuex from 'vuex'
+
+Vue.use(Vuex)
+
+
+const store = new Vuex.Store({
+  state: {
+
+  },
+  mutations:{
+    
+  }
+})
